mcp-tutorial/requirements.txt
# MCP Tutorial Dependencies
# LangChain Core
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20

# MCP Protocol
mcp>=0.1.0

# LLM Provider (Tongyi Qianwen)
dashscope>=1.14.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
